xinitrc: don't run dwm in a loop
my build of dwm can now restart if you send it SIGHUP, so no need
to we don't have to run dwm in a loop.
by not running dwm in a loop, we can kill X by just killing dwm
(whose PID is technically easier to find or whatever, who cares).